requirements.txt
loguru>=0.6.0
PyYAML>=6.0